==> f08_ampliseq_wecare_nfe_R/s05_update and_cleaning/s05_update_variant_annotations_apply_FILTER.Rmd <==
---
title: "Ampliseq wecare NFFE: update variant annotations, apply FILTER"
author: "Alexey Larionov"
date: "27 Dec 2021"
output:
  html_document:
    toc: true
    number_sections: true
    toc_float: true
editor_options:
  chunk_output_type: console
---

# Summary  

Retain only most informative fields in variants data frame  
Update types of the selected fields in variants data frame  

Overall this drops the number of variant annotations from 116 to 65  
(may be further reduced later)  

Input data:  11,483 variants x 587 samples (198 NFFE, 194 UBC & 195 CBC)  
Output data: 7,322 variants x 587 samples (198 NFFE, 194 UBC & 195 CBC)  

# Start section

<style>
pre {
  overflow-x: auto;
}
pre code {
  word-wrap: normal;
  white-space: pre;
}
</style>

```{r echo=F}
options(width = 999)
```

```{r}

# Time stamp
Sys.time()

# Clenan-up
rm(list=ls())
graphics.off()

# Memory
gc()

# Options
options(stringsAsFactors = F,
        warnPartialMatchArgs = T, 
        warnPartialMatchAttr = T, 
        warnPartialMatchDollar = T)

# Files and folders
base_folder="/Users/alexey/Documents/wecare/final_analysis_2021/reanalysis_afterNov2021/ampliseq/s02_ampliseq_wecare_nfe"
data_folder <- file.path(base_folder,"data","s05_update and_cleaning")
scripts_folder <- file.path(base_folder,"scripts","s05_update and_cleaning")
setwd(scripts_folder)

# Libraries
library(dplyr)

```

# Read data

```{r}

load(file.path(data_folder,"s04_explore_annotations_filter_by_hwe.RData"))

# Update folders
base_folder="/Users/alexey/Documents/wecare/final_analysis_2021/reanalysis_afterNov2021/ampliseq/s02_ampliseq_wecare_nfe"
data_folder <- file.path(base_folder,"data","s05_update and_cleaning")
scripts_folder <- file.path(base_folder,"scripts","s05_update and_cleaning")

```

# Check data

```{r}

# List objects
ls()

# Check sizes
dim(gt_add.mx)
dim(gt_num.mx)
dim(gt_chr.mx)

dim(alt_fraction.mx)

dim(gq.mx)
dim(dp.mx)

dim(variants.df)
dim(phenotypes.df)
table(phenotypes.df$cc, useNA = "always")

# Check consistency of rownames and colnames

sum(rownames(gt_add.mx) != rownames(gt_num.mx))
sum(rownames(gt_add.mx) != rownames(gt_chr.mx))
sum(rownames(gt_add.mx) != rownames(alt_fraction.mx))
sum(rownames(gt_add.mx) != rownames(gq.mx))
sum(rownames(gt_add.mx) != rownames(dp.mx))

sum(colnames(gt_add.mx) != colnames(gt_num.mx))
sum(colnames(gt_add.mx) != colnames(gt_chr.mx))
sum(colnames(gt_add.mx) != colnames(alt_fraction.mx))
sum(colnames(gt_add.mx) != colnames(gq.mx))
sum(colnames(gt_add.mx) != colnames(dp.mx))

sum(rownames(gt_add.mx) != rownames(variants.df))
sum(colnames(gt_add.mx) != rownames(phenotypes.df)) 

```

# Initial list of variant annotations

```{r}

colnames(variants.df)

```

# Keep only variants passed FILTER

```{r}

sort(table(variants.df$FILTER,useNA = "always"),decreasing=T)
pf_variants <- variants.df$FILTER == "PASS"
sum(pf_variants)

gt_add.mx <- gt_add.mx[pf_variants,]
gt_num.mx <- gt_num.mx[pf_variants,]
gt_chr.mx <- gt_chr.mx[pf_variants,]

alt_fraction.mx <- alt_fraction.mx[pf_variants,]

gq.mx <- gq.mx[pf_variants,]
dp.mx <- dp.mx[pf_variants,]

variants.df <- variants.df[pf_variants,]

rm(pf_variants)

```

# Keep selected annotations  

```{r}

selected_annottions=c(

"ID","CHROM","POS","REF","ALT",

"vep_SYMBOL","vep_VARIANT_CLASS","vep_Consequence","vep_IMPACT",
"vep_DISTANCE","vep_NEAREST","vep_STRAND",

"vep_Existing_variation","init_AC","init_AN","init_AF",
"QUAL","AS_VQSLOD","DP","MQ",
"NEGATIVE_TRAIN_SITE","POSITIVE_TRAIN_SITE",
"hwe_biallelic","InbreedingCoeff","ExcessHet",
"Multiallelic",

"vep_Allele",
"vep_Codons","vep_Amino_acids","vep_DOMAINS",
"vep_cDNA_position","vep_CCDS","vep_CDS_position","vep_Protein_position",

"SIFT_call","SIFT_score","PolyPhen_call","PolyPhen_score",

"vep_MAX_AF","vep_MAX_AF_POPS",

"vep_gnomAD_AF","vep_gnomAD_AFR_AF","vep_gnomAD_AMR_AF",
"vep_gnomAD_ASJ_AF","vep_gnomAD_EAS_AF","vep_gnomAD_FIN_AF",
"vep_gnomAD_NFE_AF","vep_gnomAD_OTH_AF","vep_gnomAD_SAS_AF",

"vep_CADD_RAW","vep_CADD_PHRED",

"CLNSIG","CLNREVSTAT","CLNDN","CLNSIGCONF","CLNVC","CLNVCSO","CLNVI",
"GENEINFO","MC","ORIGIN","RS","ALLELEID","CLNDISDB","CLNHGVS")

length(selected_annottions)
variants.df <- variants.df[,selected_annottions]
dim(variants.df)

# Clean-up
rm(selected_annottions)

```

# Update types of retained annotations in variants.df  

```{r}

str(variants.df, list.len=Inf)

variants.df$init_AC <- as.integer(variants.df$init_AC)
variants.df$init_AF <- as.numeric(variants.df$init_AF)

variants.df$AS_VQSLOD <- as.numeric(variants.df$AS_VQSLOD)

variants.df$vep_DISTANCE <- as.integer(variants.df$vep_DISTANCE)
variants.df$vep_STRAND <- as.integer(variants.df$vep_STRAND)

variants.df$vep_MAX_AF <- as.numeric(variants.df$vep_MAX_AF)

variants.df$vep_gnomAD_AF <- as.numeric(variants.df$vep_gnomAD_AF)
variants.df$vep_gnomAD_AFR_AF <- as.numeric(variants.df$vep_gnomAD_AFR_AF)
variants.df$vep_gnomAD_AMR_AF <- as.numeric(variants.df$vep_gnomAD_AMR_AF)
variants.df$vep_gnomAD_ASJ_AF <- as.numeric(variants.df$vep_gnomAD_ASJ_AF)
variants.df$vep_gnomAD_EAS_AF <- as.numeric(variants.df$vep_gnomAD_EAS_AF)
variants.df$vep_gnomAD_FIN_AF <- as.numeric(variants.df$vep_gnomAD_FIN_AF)
variants.df$vep_gnomAD_NFE_AF <- as.numeric(variants.df$vep_gnomAD_NFE_AF)
variants.df$vep_gnomAD_OTH_AF <- as.numeric(variants.df$vep_gnomAD_OTH_AF)
variants.df$vep_gnomAD_SAS_AF <- as.numeric(variants.df$vep_gnomAD_SAS_AF)

variants.df$vep_CADD_RAW <- as.numeric(variants.df$vep_CADD_RAW)
variants.df$vep_CADD_PHRED <- as.numeric(variants.df$vep_CADD_PHRED)

str(variants.df, list.len=Inf)

```

# Check data

```{r}

# List objects
ls()

# Check sizes
dim(gt_add.mx)
dim(gt_num.mx)
dim(gt_chr.mx)

dim(alt_fraction.mx)

dim(gq.mx)
dim(dp.mx)

dim(variants.df)
dim(phenotypes.df)
table(phenotypes.df$cc, useNA = "always")

# Check consistency of rownames and colnames

sum(rownames(gt_add.mx) != rownames(gt_num.mx))
sum(rownames(gt_add.mx) != rownames(gt_chr.mx))
sum(rownames(gt_add.mx) != rownames(alt_fraction.mx))
sum(rownames(gt_add.mx) != rownames(gq.mx))
sum(rownames(gt_add.mx) != rownames(dp.mx))

sum(colnames(gt_add.mx) != colnames(gt_num.mx))
sum(colnames(gt_add.mx) != colnames(gt_chr.mx))
sum(colnames(gt_add.mx) != colnames(alt_fraction.mx))
sum(colnames(gt_add.mx) != colnames(gq.mx))
sum(colnames(gt_add.mx) != colnames(dp.mx))

sum(rownames(gt_add.mx) != rownames(variants.df))
sum(colnames(gt_add.mx) != rownames(phenotypes.df)) 

```

# Save result

```{r}

save.image(file.path(data_folder, "s05_update_variant_annotations_apply_FILTER.RData"))

```

# Final section

```{r}

ls()
sessionInfo()
Sys.time()
gc()

```
